+++ jest.setup.js
@@ -1,0 +1,6 @@
+// jest.setup.js
+/* eslint-disable import/no-extraneous-dependencies */
+import { configure } from 'enzyme'
+import Adapter from 'enzyme-adapter-react-16'
+
+configure({ adapter: new Adapter() })