+++ packages/webpack-box/lib/index.js
@@ -1,6 +1,6 @@
 function cleanArgs(cmd) {
   const args = {}
-  cmd.options.forEach(o => {
+  cmd.options && cmd.options.forEach(o => {
     const key = camelize(o.long.replace(/^--/, ''))
     if (typeof cmd[key] !== 'function' && typeof cmd[key] !== 'undefined') {
       args[key] = cmd[key]