+++ CHANGELOG.md
@@ -1,9 +1,28 @@
 ## [1.0.9](https://github.com/luoxue-victor/learn_webpack/compare/v1.2.12...v1.0.9) (2020-01-11)
 
+### 🌟 新功能
+范围|描述|commitId
+--|--|--
+ / | 增加jest单元测试 | [9ecaee8](https://github.com/luoxue-victor/learn_webpack/commit/9ecaee8), closes [#9](https://github.com/luoxue-victor/learn_webpack/issues/9)
+ packages/react | Sentry错误监控 | [b417e40](https://github.com/luoxue-victor/learn_webpack/commit/b417e40), closes [#22](https://github.com/luoxue-victor/learn_webpack/issues/22)
+
+
 ### 📝 文档
 范围|描述|commitId
 --|--|--
  changlog | 修改changelog | [4a8da45](https://github.com/luoxue-victor/learn_webpack/commit/4a8da45)
+
+
+### 🚀 性能优化
+范围|描述|commitId
+--|--|--
+ / | 更改jest单元测试配置 | [4aadd64](https://github.com/luoxue-victor/learn_webpack/commit/4aadd64)
+
+
+### chore
+范围|描述|commitId
+--|--|--
+ 合并代码 | [dfdc68f](https://github.com/luoxue-victor/learn_webpack/commit/dfdc68f)
 
 ## [1.2.12](https://github.com/luoxue-victor/learn_webpack/compare/v1.2.11...v1.2.12) (2019-12-27)